requirements.txt
--index-url https://download.pytorch.org/whl/rocm6.4
--extra-index-url https://pypi.amd.com/simple
--extra-index-url https://pypi.amd.com/rocm-6.4.3/simple
ITK
albumentations
amd-cupy
amd-hipcim
amd-monai
black>=25.1.0
blinker
clearml>=1.10.0rc0
click
coverage>=5.5
cython==3.0.12
einops
filelock<3.12.0  # https://github.com/microsoft/nni/issues/5523
fire
flake8-bugbear<=24.2.6  # https://github.com/Project-MONAI/MONAI/issues/7690
flake8-comprehensions
flake8>=3.8.1
gdown>=4.7.3
git+https://github.com/Project-MONAI/MetricsReloaded@monai-support#egg=MetricsReloaded
git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
h5py
huggingface_hub
ignite
imagecodecs>=2021.6.8; platform_system == "Linux" or platform_system == "Darwin"
ipython
isort>=5.1, <6.0
itk>=5.2
jsonschema
lazy-loader>=0.4
lmdb
lpips==0.1.4
matplotlib>=3.7
mccabe
mlflow>=2.12.2
moviepy # Optional
mypy>=1.5.0, <1.12.0
nbsphinx
nibabel
ninja
nni==2.10.1; platform_system == "Linux" and "arm" not in platform_machine and "aarch" not in platform_machine
numpy>=1.24,<3.0
numpydoc>=1.7
onnx>=1.13.0
onnx_graphsurgeon
onnxruntime; python_version <= '3.10'
opencv-contrib-python # Optional
opencv-python-headless>=4.6
openslide-bin
openslide-python>=1.3.0
optuna  				# 4.3.0
packaging				# 25.0
pandas
parameterized
pep8-naming
pillow!=8.3.0  # https://github.com/python-pillow/Pillow/issues/5571
pip
plotly
polygraphy
pooch>=1.6.0
pre-commit
psutil>=5.8.0
pyamg>=5.0.0
pycodestyle
pydata-sphinx-theme
pydicom
pyflakes
pynrrd
pytest-cov>=2.12.1
pytest-lazy-fixtures>=1.0.0
pytest-xdist
pytest>=7.0.0,<9.0.0a0
pytorch-ignite 
pytype>=2020.6.1; platform_system != "Windows"
pyvips
pywavelets>=1.6
pyyaml
recommonmark
redis                              	# 6.2.0
requests
ruff
scikit-image>=0.19.0,<0.26.0a0
scipy>=1.12.0; python_version >= '3.9'
setuptools>=70.2.0,<=79.0.1; python_version >= "3.12"
sphinx>=8.0.0,<8.2.0
streamlit
streamlit-option-menu
tensorboard>=2.12.0  # https://github.com/Project-MONAI/MONAI/issues/7434
tensorboardX
tifffile>=2022.8.12; platform_system == "Linux" or platform_system == "Darwin"
torch==2.8.0
torchio
tqdm>=4.47.0
transformers>=4.36.0, <4.41.0; python_version <= '3.10'
typeguard<3  # https://github.com/microsoft/nni/issues/5457
types-PyYAML
types-setuptools
wheel 
zarr
